moved Loaders to UI
